+++ crates/types/src/services/block_producer.rs
@@ -1,8 +1,15 @@
 //! Types related to block producer service.
 
 use crate::{
-    blockchain::header::PartialBlockHeader,
+    blockchain::{
+        block::Block,
+        header::{
+            PartialBlockHeader,
+            StateTransitionBytecodeVersion,
+        },
+    },
     fuel_tx::ContractId,
+    services::executor::ExecutionTypes,
 };
 
 /// The components required to produce a block.
@@ -19,4 +26,25 @@ pub struct Components<Source> {
     pub coinbase_recipient: ContractId,
     /// The gas price for all transactions in the block.
     pub gas_price: u64,
+}
+
+impl<TxSource> ExecutionTypes<Components<TxSource>, Block> {
+    /// Returns the state transition bytecode version of the block.
+    pub fn state_transition_version(&self) -> StateTransitionBytecodeVersion {
+        match self {
+            ExecutionTypes::DryRun(component) => {
+                component
+                    .header_to_produce
+                    .state_transition_bytecode_version
+            }
+            ExecutionTypes::Production(component) => {
+                component
+                    .header_to_produce
+                    .state_transition_bytecode_version
+            }
+            ExecutionTypes::Validation(block) => {
+                block.header().state_transition_bytecode_version
+            }
+        }
+    }
 }